a little refactoring
Refactor: Standardize icon usage and simplify theming
Replaces custom icon wrappers with a new unified `Icon` component that uses `lucide-react-native` directly. This change simplifies icon implementation and styling across the application by leveraging Tailwind CSS classes instead of hardcoded props.

Removes the custom `useColorScheme` hook and manual dark mode logic in favor of a centralized theming system using semantic CSS variables, streamlining component styling.

Additionally, this update:
- Simplifies the settings page by removing the local profile data cache and the dark mode toggle UI.
- Moves the `AttendanceCalendar` component into a more structured directory.
- Deletes the unused `Table` component and various icon wrapper files.
