--- lib/icons/AlertCircle.tsx
@@ -1,4 +1,0 @@
-import { AlertCircle } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(AlertCircle);
-export { AlertCircle };
