--- lib/icons/Bell.tsx
@@ -1,4 +1,0 @@
-import { Bell } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(Bell);
-export { Bell };
