--- lib/icons/Calendar.tsx
@@ -1,4 +1,0 @@
-import { Calendar } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(Calendar);
-export { Calendar };
