--- lib/icons/Camera.tsx
@@ -1,4 +1,0 @@
-import { Camera } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(Camera);
-export { Camera };
