--- lib/icons/CameraOff.tsx
@@ -1,4 +1,0 @@
-import { CameraOff } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(CameraOff);
-export { CameraOff };
