--- lib/icons/CameraSwitch.tsx
@@ -1,4 +1,0 @@
-import { RefreshCw } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(RefreshCw);
-export { RefreshCw as CameraSwitch }; // CameraSwitch icon does not exist in lucide, using RefreshCw as substitute
