--- lib/icons/CheckCircle.tsx
@@ -1,4 +1,0 @@
-import { CheckCircle } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(CheckCircle);
-export { CheckCircle };
