--- lib/icons/ChevronDown.tsx
@@ -1,5 +1,0 @@
-import { ChevronDown } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-
-iconWithClassName(ChevronDown);
-export { ChevronDown };
