--- lib/icons/ChevronLeft.tsx
@@ -1,4 +1,0 @@
-import { ChevronLeft } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(ChevronLeft);
-export { ChevronLeft };
