--- lib/icons/ChevronRight.tsx
@@ -1,4 +1,0 @@
-import { ChevronRight } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(ChevronRight);
-export { ChevronRight };
