--- lib/icons/ClipboardPenLine.tsx
@@ -1,4 +1,0 @@
-import { ClipboardPenLine } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(ClipboardPenLine);
-export { ClipboardPenLine };
