--- lib/icons/Clock.tsx
@@ -1,4 +1,0 @@
-import { Clock } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(Clock);
-export { Clock };
