--- lib/icons/Eye.tsx
@@ -1,5 +1,0 @@
-import { Eye } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-
-iconWithClassName(Eye);
-export { Eye };
