--- lib/icons/EyeOff.tsx
@@ -1,5 +1,0 @@
-import { EyeOff } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-
-iconWithClassName(EyeOff);
-export { EyeOff };
