--- lib/icons/FileText.tsx
@@ -1,4 +1,0 @@
-import { FileText } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(FileText);
-export { FileText };
