--- lib/icons/HelpCircle.tsx
@@ -1,4 +1,0 @@
-import { HelpCircle } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(HelpCircle);
-export { HelpCircle };
