--- lib/icons/History.tsx
@@ -1,4 +1,0 @@
-import { History } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(History);
-export { History };
