--- lib/icons/Key.tsx
@@ -1,4 +1,0 @@
-import { Key } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(Key);
-export { Key };
