--- lib/icons/Loader2.tsx
@@ -1,4 +1,0 @@
-import { Loader2 } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(Loader2);
-export { Loader2 };
