--- lib/icons/LockIcon.tsx
@@ -1,4 +1,0 @@
-import { LockIcon } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(LockIcon);
-export { LockIcon };
