--- lib/icons/LogOut.tsx
@@ -1,4 +1,0 @@
-import { LogOut } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(LogOut);
-export { LogOut };
