--- lib/icons/MapPin.tsx
@@ -1,4 +1,0 @@
-import { MapPin } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(MapPin);
-export { MapPin };
