--- lib/icons/MapPinOff.tsx
@@ -1,4 +1,0 @@
-import { MapPinOff } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(MapPinOff);
-export { MapPinOff };
