--- lib/icons/Moon.tsx
@@ -1,4 +1,0 @@
-import { Moon } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(Moon);
-export { Moon };
