--- lib/icons/MousePointerClick.tsx
@@ -1,4 +1,0 @@
-import { MousePointerClick } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(MousePointerClick);
-export { MousePointerClick };
