--- lib/icons/RefreshCw.tsx
@@ -1,4 +1,0 @@
-import { RefreshCw } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(RefreshCw);
-export { RefreshCw };
