--- lib/icons/Settings.tsx
@@ -1,4 +1,0 @@
-import { Settings } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(Settings);
-export { Settings };
