--- lib/icons/User.tsx
@@ -1,4 +1,0 @@
-import { User } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(User);
-export { User };
