--- lib/icons/UserCheck.tsx
@@ -1,4 +1,0 @@
-import { UserCheck } from "lucide-react-native";
-import { iconWithClassName } from "./iconWithClassName";
-iconWithClassName(UserCheck);
-export { UserCheck };
