--- lib/icons/iconWithClassName.ts
@@ -1,17 +1,0 @@
-import type { LucideIcon } from "lucide-react-native";
-import { cssInterop } from "nativewind";
-
-export function iconWithClassName(icon: LucideIcon) {
-  // Make sure the icon component can accept className props through nativewind
-  cssInterop(icon, {
-    className: {
-      target: "style",
-      nativeStyleToProp: {
-        color: true,
-        opacity: true,
-        width: true,
-        height: true,
-      },
-    },
-  });
-}
