--- lib/useColorScheme.tsx
@@ -1,12 +1,0 @@
-import { useColorScheme as useNativewindColorScheme } from "nativewind";
-
-export function useColorScheme() {
-  const { colorScheme, setColorScheme, toggleColorScheme } =
-    useNativewindColorScheme();
-  return {
-    colorScheme: colorScheme ?? "dark",
-    isDarkColorScheme: colorScheme === "dark",
-    setColorScheme,
-    toggleColorScheme,
-  };
-}
